nmstate: another try for enablement

+++ nmcli/features/environment.py
@@ -1341,17 +1341,13 @@ def before_scenario(context, scenario):
                 reload_NM_service()
 
             if 'nmstate_setup' in scenario.tags:
-                # Set veths as managed if we don't use veths yet
-                manage_veths ()
+                # Skip on deployments where we do not have veths
+                if not os.path.isfile('/tmp/nm_newveth_configured'):
+                    sys.exit(77)
 
-                # Rename eth1 and eth2 to nmstate_eth1 and nmstate_eth2
-                # as we need new eth1 and eth2 to be free
-                call("ip link set dev eth1 down", shell=True)
-                call("ip link set dev eth1 name nmstate_eth1", shell=True)
-                call("ip link set dev nmstate_eth1 up", shell=True)
-                call("ip link set dev eth2 down", shell=True)
-                call("ip link set dev eth2 name nmstate_eth2", shell=True)
-                call("ip link set dev nmstate_eth2 up", shell=True)
+                call("sh prepare/vethsetup.sh teardown", shell=True)
+
+                manage_veths ()
 
                 # Install some deps
                 call("yum -y install python3-devel rpm-build", shell=True)
@@ -1705,35 +1701,17 @@ def after_scenario(context, scenario):
                 call("ip link del eth1", shell=True)
                 call("ip link del eth2", shell=True)
 
-                call("ip link set dev nmstate_eth1 down", shell=True)
-                call("ip link set dev nmstate_eth1 name eth1", shell=True)
-                call("ip link set dev eth1 up", shell=True)
-                call("ip link set dev nmstate_eth2 down", shell=True)
-                call("ip link set dev nmstate_eth2 name eth2", shell=True)
-                call("ip link set dev eth2 up", shell=True)
-                reset_hwaddr_nmcli('eth1')
-                reset_hwaddr_nmcli('eth2')
-
-                call("nmcli con del eth1 eth2 linux-br0 dhcpcli dhcpsrv bond99", shell=True)
+                call("nmcli con del eth1 eth2 linux-br0 dhcpcli dhcpsrv bond99 eth1.101 eth1.102", shell=True)
                 call("nmcli device delete dhcpsrv", shell=True)
                 call("nmcli device delete dhcpcli", shell=True)
                 call("nmcli device delete bond99", shell=True)
-
-                call("nmcli con up testeth1 && nmcli con down testeth1", shell=True)
-                call("nmcli con up testeth2 && nmcli con down testeth2", shell=True)
 
                 # in case of fail we need to kill this
                 call('rm -rf /etc/dnsmasq.d/nmstate.conf', shell=True)
                 call('systemctl stop dnsmasq', shell=True)
 
-                if not os.path.isfile('/tmp/nm_newveth_configured'):
-                    # Undo: set veths as managed if we don't use veths yet
-                    unmanage_veths ()
-                else:
-                    call('sh prepare/vethsetup.sh check', shell=True)
+                call('sh prepare/vethsetup.sh check', shell=True)
 
-                # Restore all connections back as before
-                restore_connections ()
                 wait_for_testeth0 ()
 
                 print("* attaching nmstate log")
